Fix soil moisture supersaturation triggered by unit conversion.
Conversion of JULES soil moisture from kg m^-2 to m^3 m^-3 can lead to slightly
oversaturated values if the values in the original units are already at
saturation.  This is due to floating point rounding error.  The added code
resets the converted soil moistures to the saturated value in that case.

(The values are only about 1e-7 m^3 m^-3 over the saturation limit, so it's
not a huge deal, but it seems best to fix it immediately.)